add embeddable address class
