+++ src/main/java/com/heck/auth/api/models/records/User.java
@@ -1,0 +1,25 @@
+package com.heck.auth.api.models.records;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private long id;
+
+    private String firstName;
+    private String lastName;
+    private String email;
+    private String password;
+    private String avatar;
+}
